+++ tests/cli.rs
@@ -18,6 +18,14 @@ fn test_index_cmd_success_uniprot() -> Result<(), Box<dyn std::error::Error>> {
 }
 
 #[test]
+fn test_max_length_cmd_success() -> Result<(), Box<dyn std::error::Error>> {
+    let mut cmd = Command::cargo_bin("fastatools")?;
+    cmd.arg("max-length").arg("resources/test.fasta");
+    cmd.assert().stderr(predicate::str::contains("506"));
+    Ok(())
+}
+
+#[test]
 fn test_index_cmd_success_id_only() -> Result<(), Box<dyn std::error::Error>> {
     let mut cmd = Command::cargo_bin("fastatools")?;
     cmd.arg("index").arg("resources/test_short_desc.fasta");