+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleasedd
+
+### Added
+
+- fastatools max-length: returns maximal sequence length in a fasta file
+
 ## [0.1.1]
 
 ### Fixed